Added search by name

+++ src/main/java/com/enciu/rohighwaypulse/module/Segment.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Data
 public class Segment {
     private String name;
-    private Integer length;
+    private Float length;
     private String status;
     private String contractor;
     private Integer estimatedCompletionYear;
